embedding_service/requirements.txt
fastapi>=0.123.10
uvicorn>=0.40.0
torch>=2.9.1 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.57.6
sentence-transformers>=5.2.0
pillow>=12.1.0
pydantic>=2.12.5
python-dotenv>=1.2.1
requests>=2.32.4